# Fix type instability of `entropy` and generalize `crossentropy` and `kldivergence`