--- a/test/REQUIRE
+++ b/test/REQUIRE
@@ -1,2 +1,0 @@
-DataFrames
-GLM