While I was working on a PR to Distributions, I noticed that `entropy` is not type stable:
```julia
julia> @inferred entropy([0, 1])
ERROR: return type Float64 does not match inferred return type Union{Float64, Int64}
```
This PR fixes the issue by using `LogExpFunctions.xlogx` instead of `x -> x > 0 ? zero(x) : z * log(x)`. Moreover, it bases `crossentropy` and `kldivergence` on `LogExpFunctions.xlogy` and generalizes them to non-Float64 return types and arguments with different element types.