@nalimilan 2021-09-04
Thanks, that indeed simplifies the code a lot. Have you checked that performance doesn't regress? Also, does it still work for empty inputs?