@devmotion 2021-09-05
> Thanks, that indeed simplifies the code a lot. Have you checked that performance doesn't regress?
> Also, does it still work for empty inputs?

With the latest release:
```julia
julia> using StatsBase, BenchmarkTools, Random

julia> Random.seed!(1234);

julia> p, q = rand(100_000), rand(100_000);

julia> @btime entropy($p);
  753.493 μs (0 allocations: 0 bytes)

julia> @btime crossentropy($p, $q);
  779.563 μs (0 allocations: 0 bytes)

julia> @btime kldivergence($p, $q);
  1.540 ms (0 allocations: 0 bytes)

julia> pint, qint = rand((0, 1), 100_000), rand((0, 1), 100_000);

julia> @btime entropy($pint);
  819.110 μs (0 allocations: 0 bytes)

julia> @btime crossentropy($pint, $qint);
  910.370 μs (0 allocations: 0 bytes)

julia> @btime kldivergence($pint, $qint);
  948.375 μs (0 allocations: 0 bytes)

julia> entropy(Float64[])
ERROR: ArgumentError: reducing over an empty collection is not allowed
Stacktrace:
...

julia> crossentropy(Float64[], Float64[])
-0.0

julia> kldivergence(Float64[], Float64[])
0.0
```

With this PR:
```julia
julia> using StatsBase, BenchmarkTools, Random

julia> Random.seed!(1234);

julia> p, q = rand(100_000), rand(100_000);

julia> @btime entropy($p);
  837.509 μs (0 allocations: 0 bytes)

julia> @btime crossentropy($p, $q);
  1.133 ms (0 allocations: 0 bytes)

julia> @btime kldivergence($p, $q);
  1.680 ms (0 allocations: 0 bytes)

julia> pint, qint = rand((0, 1), 100_000), rand((0, 1), 100_000);

julia> @btime entropy($pint);
  1.116 ms (0 allocations: 0 bytes)

julia> @btime crossentropy($pint, $qint);
  1.210 ms (0 allocations: 0 bytes)

julia> @btime kldivergence($pint, $qint);
  1.725 ms (0 allocations: 0 bytes)

julia> entropy(Float64[])
ERROR: ArgumentError: reducing over an empty collection is not allowed
Stacktrace:
...

julia> crossentropy(Float64[], Float64[])
ERROR: ArgumentError: reducing over an empty collection is not allowed
Stacktrace:
...

julia> kldivergence(Float64[], Float64[])
ERROR: ArgumentError: reducing over an empty collection is not allowed
Stacktrace:
...
```

So there seems to be some performance regression and empty inputs don't work but also don't work with `entropy` in the latest release. I will look into why there is a performance drop. For `entropy` I just updated the mapping function in `sum` with `xlogx`, so I wonder why it is slower than before - maybe it could be caused by the use of `ifelse` in LogExpFunctions? If empty inputs should be supported, I could just add a check and compute the return type with `zero(eltype(p))` etc. I assume (this would also fix the problem in the latest release I guess).