@devmotion 2021-09-05
On a second thought, it is not completely clear to me anymore if it is reasonable to define `entropy`, `kldivergence` or `crossentropy` for empty collections, and what values one would expect.