@nalimilan 2021-09-06
Given that these are sums, isn't it logical to return zero for empty inputs? Anyway, throwing an error instead would be breaking so we would at least need to keep returning zero, possible with a deprecation warning.