@devmotion 2021-09-06
`entropy` already throws an error but indeed it would be a breaking change for `crossentropy` and `kldivergence`.

My main motivation for throwing an error would be that empty vectors don't represent probability distributions.