@devmotion 2021-10-01
So what's the plan here? Return a typestable 0 for empty inputs also for `entropy`, as currently done for `crossentropy` and `kldivergence`? Or keep throwing a (possibly more meaningful) error in the case of `entropy`?