@mschauer 2021-10-01
> My main motivation for throwing an error would be that empty vectors don't represent probability distributions.

I find that convincing.