@nalimilan 2021-10-02
As you prefer, but I just think we shouldn't add new errors for now to avoid being breaking. A deprecation warning would be OK, with a PR to turn them into errors in the next breaking release.

BTW should we check that the values sum to one if we want to ensure they are probability distributions (with a way to skip the check)?