@devmotion 2021-10-03
I reran the benchmarks with Julia 1.6.3 and also compared it with the `sum(Broadcast....)` suggestion. There is still a performance regression (switching from `ifelse` to ternary operators in LogExpFunctions helped a bit but did not resolve them) and it seems the comparison between the PR and `sum(Broadcast...)` is not conclusive and none of the approaches is clearly better than the other.

Without this PR:
```julia
julia> using StatsBase, BenchmarkTools, Random

julia> Random.seed!(1234);

julia> p, q = rand(100_000), rand(100_000);

julia> @btime entropy($p);
  732.629 μs (0 allocations: 0 bytes)

julia> @btime crossentropy($p, $q);
  766.728 μs (0 allocations: 0 bytes)

julia> @btime kldivergence($p, $q);
  1.441 ms (0 allocations: 0 bytes)

julia> pint, qint = rand((0, 1), 100_000), rand((0, 1), 100_000);

julia> @btime entropy($pint);
  788.951 μs (0 allocations: 0 bytes)

julia> @btime crossentropy($pint, $qint);
  895.965 μs (0 allocations: 0 bytes)

julia> @btime kldivergence($pint, $qint);
  919.979 μs (0 allocations: 0 bytes)
```

With this PR (commit https://github.com/JuliaStats/StatsBase.jl/pull/714/commits/97d3bfadae8d560c6d0aeda799cedcaffeaf677a):
```julia
julia> using StatsBase, BenchmarkTools, Random

julia> Random.seed!(1234);

julia> p, q = rand(100_000), rand(100_000);

julia> @btime entropy($p);
  763.036 μs (0 allocations: 0 bytes)

julia> @btime crossentropy($p, $q);
  1.100 ms (0 allocations: 0 bytes)

julia> @btime kldivergence($p, $q);
  1.557 ms (0 allocations: 0 bytes)

julia> pint, qint = rand((0, 1), 100_000), rand((0, 1), 100_000);

julia> @btime entropy($pint);
  996.637 μs (0 allocations: 0 bytes)

julia> @btime crossentropy($pint, $qint);
  1.178 ms (0 allocations: 0 bytes)

julia> @btime kldivergence($pint, $qint);
  1.604 ms (0 allocations: 0 bytes)
```

With `sum(Broadcast....)`:
```julia
julia> using StatsBase, BenchmarkTools, Random

julia> Random.seed!(1234);

julia> p, q = rand(100_000), rand(100_000);

julia> @btime entropy($p);
  763.060 μs (0 allocations: 0 bytes)

julia> @btime crossentropy($p, $q);
  899.686 μs (0 allocations: 0 bytes)

julia> @btime kldivergence($p, $q);
  1.582 ms (0 allocations: 0 bytes)

julia> pint, qint = rand((0, 1), 100_000), rand((0, 1), 100_000);

julia> @btime entropy($pint);
  997.401 μs (0 allocations: 0 bytes)

julia> @btime crossentropy($pint, $qint);
  1.312 ms (0 allocations: 0 bytes)

julia> @btime kldivergence($pint, $qint);
  1.666 ms (0 allocations: 0 bytes)
```