@devmotion 2021-10-05
I updated the PR, it handles (and deprecates) empty collections of probabilities and uses pairwise summation now for `crossentropy` and `kldivergence`.